+++ src/main/java/ru/iimm/ontology/visualization/ui/mvp/ModelCFrameOntology.java
@@ -10,7 +10,7 @@ import ru.iimm.ontology.cftools.CFrameOnt;
 import ru.iimm.ontology.visualization.tools.CFrameDecorator;
 
 /**
- *
+ * Модель предстовляющий отологию с КФ.
  * @author Danilov
  * @version 0.1
  */
@@ -18,28 +18,30 @@ public interface ModelCFrameOntology
 {
 	
 	/**
-	 * @return
+	 * Полулчить список с CFrame-обортками.
 	 */
 	ArrayList<CFrameDecorator> getCframes();
 
 	/**
-	 * @param o1
-	 * @return
+	 * Получить ярлык CFrame.
+	 * @param cframe - CFrame.
+	 * @return - Ярлык в виде строки.
 	 */
 	String getLabel(CFrame cframe);
-	
+	/**Получить онтологию с КФ.*/
 	CFrameOnt getOntology();
+	/**Задать онтологию с КФ.*/
 	void setOntology(CFrameOnt ont);
 
 	/**
-	 * @param owlName
-	 * @return
+	 * Получить ярлык OWLNamedIndividual.
+	 * @param owlName - OWLNamedIndividual.
+	 * @return - Ярлык в виде строки.
 	 */
 	String getLabel(OWLNamedIndividual owlName);
 
 	/**
-	 * @param iri
-	 * @return
+	 * Получить анотацию у IRI.
 	 */
 	String getAnnotationValue(IRI iri);
 }