+++ src/main/java/ru/iimm/ontology/visualization/ui/mvp/PresenterCFrameGSVisitor.java
@@ -3,7 +3,7 @@ package ru.iimm.ontology.visualization.ui.mvp;
 import ru.iimm.ontology.visualization.tools.VisualMethodVisitorInt;
 
 /**
- *
+ * Presenter-Visitor для визуализации фрейма библиотекой Graph Stream. 
  * @author Danilov
  * @version 0.1
  */